Use ODK 1.6.
The repository had been updated to the latest development version of the
ODK, using the new ODK-Core system. This is great from the point of view
of a ODK developer (to test the latest version), but not from the point
of view of an ontology maintainer (since it means in particular that we
can't run the test suite on GitHub).

I am working on FBbi with my ontology maintainer hat, _not_ my ODK
developer hat, so we better use the latest _released_ version of the ODK.