Update metadata.
Update metadata to reflect the new maintainers of the ontology. For now,
this is assuming the ontology will remain in my own GitHub space, which
will _not_ be the case for long -- but making those changes now will
make it easier to later update the metadata as we will just have to
replace the name of my GitHub account with the name of whichever
organisation will host FBbi.

For the ontology metadata, we use dcterms annotations rather than
rdfs:comment, and we include known contributors for which we have an
ORCID.